fix: add missing reward types and fix bigint type assertions
- Added signup_bonus, signup_commission_l1/l2/l3 to typeLabels Record
- Fixed bigint type assertions for readContract returns in token-transfer-service
- Ensures TypeScript strict mode passes for Vercel build

Made by mbxarts.com The Moon in a Box property

Co-Author: Godez22

+++ lib/web3/token-transfer-service.ts
@@ -151,7 +151,7 @@ export async function getDeployerCGCBalance(): Promise<{
     abi: ERC20_ABI,
     functionName: 'balanceOf',
     args: [deployerAddress],
-  });
+  }) as bigint;
 
   const balanceFormatted = formatUnits(balance, CGC_DECIMALS);
   const balanceNumber = parseFloat(balanceFormatted);
@@ -236,7 +236,7 @@ export async function transferCGC(
       abi: ERC20_ABI,
       functionName: 'balanceOf',
       args: [walletClient.account.address],
-    });
+    }) as bigint;
 
     if (deployerBalance < amountWei) {
       console.error(`[TokenTransfer] Insufficient CGC balance. Need ${amount}, have ${formatUnits(deployerBalance, CGC_DECIMALS)}`);
